+++ clock-utils/src/main/java/org/clock/utils/dao/impl/CommonDAOImpl.java
@@ -37,7 +37,7 @@ public class CommonDAOImpl implements ICommonDAO{
 	}
 
 	@Override
-	public <T>T findById(Class<T> classType, long id) throws Exception {
+	public <T>T findById(Class<T> classType, int id) throws Exception {
 		return em.find(classType,id);
 	}
